requirements-dev.txt
fastapi>=0.100
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv==0.21.1
python-jose==3.3.0
PyYAML==6.0
starlette>=0.24
toml==0.10.2
uvicorn==0.18.3
streamlit==1.16.0
requests==2.27.1
streamlit-nested-layout==0.1.1
celery==5.2.7
fastapi-mail[aioredis]==1.4.1
pymongo==4.4.1
motor>=3.1.1
passlib>=1.7.4